requirements.txt
Pillow
django
djangorestframework
django-cool
